requirements.txt
fastapi==0.111.0
httpx==0.27.0
apscheduler==3.10.4
uvicorn==0.30.1
python-dotenv==1.0.1
